+++ tests/parser/test_openmp.py
@@ -1,8 +1,9 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 # coding: utf-8
-
+import pytest
 from pyccel.parser.syntax.openmp import parse
 
+@pytest.mark.language_agnostic
 def test_parallel():
     d = parse(stmts='#$ omp parallel private(idx)')