+++ CHANGELOG.md
@@ -67,6 +67,7 @@ All notable changes to this project will be documented in this file.
 -   #2484 : Fix optional `TypeVar` arguments.
 -   #2406 : Fix bug when returning an expression that is translated to multiple lines.
 -   #2292 : Fix memory leak when memory is allocated in a function.
+-   #2356 : Fix missing language markers in tests.
 -   Rename `main` function when translating to C.
 
 ### Changed